Lots'n lotsa renaming to effect the renaming of DataSourceDownloads model to Downloads and creating bridge table
