+++ app/helpers/data_source_downloads_helper.rb
@@ -1,2 +1,2 @@
-module DataSourceDownloadsHelper
+module DownloadsHelper
 end
