--- app/views/data_source_downloads/destroy.html.erb
@@ -1,2 +1,0 @@
-<h1>DataSourceDownloads#destroy</h1>
-<p>Find me in app/views/data_source_downloads/destroy.html.erb</p>
