--- app/views/data_source_downloads/download.html.erb
@@ -1,2 +1,0 @@
-<h1>DataSourceDownloads#download</h1>
-<p>Find me in app/views/data_source_downloads/download.html.erb</p>
