--- app/views/data_source_downloads/test_post_params.html.erb
@@ -1,2 +1,0 @@
-<h1>DataSourceDownloads#test_post_params</h1>
-<p>Find me in app/views/data_source_downloads/test_post_params.html.erb</p>
