--- app/views/data_source_downloads/test_save_path.html.erb
@@ -1,2 +1,0 @@
-<h1>DataSourceDownloads#test_save_path</h1>
-<p>Find me in app/views/data_source_downloads/test_save_path.html.erb</p>
