--- app/views/data_source_downloads/test_url.html.erb
@@ -1,2 +1,0 @@
-<h1>DataSourceDownloads#test_url</h1>
-<p>Find me in app/views/data_source_downloads/test_url.html.erb</p>
