--- app/views/data_source_downloads/update.html.erb
@@ -1,2 +1,0 @@
-<h1>DataSourceDownloads#update</h1>
-<p>Find me in app/views/data_source_downloads/update.html.erb</p>
