+++ app/views/downloads/destroy.html.erb
@@ -1,0 +1,2 @@
+<h1>Downloads#destroy</h1>
+<p>Find me in app/views/downloads/destroy.html.erb</p>
