+++ app/views/downloads/download.html.erb
@@ -1,0 +1,2 @@
+<h1>Downloads#download</h1>
+<p>Find me in app/views/downloads/download.html.erb</p>
