+++ app/views/downloads/test_post_params.html.erb
@@ -1,0 +1,2 @@
+<h1>Downloads#test_post_params</h1>
+<p>Find me in app/views/downloads/test_post_params.html.erb</p>
