+++ app/views/downloads/test_save_path.html.erb
@@ -1,0 +1,2 @@
+<h1>Downloads#test_save_path</h1>
+<p>Find me in app/views/downloads/test_save_path.html.erb</p>
