+++ app/views/downloads/test_url.html.erb
@@ -1,0 +1,2 @@
+<h1>Downloads#test_url</h1>
+<p>Find me in app/views/downloads/test_url.html.erb</p>
