+++ app/views/downloads/update.html.erb
@@ -1,0 +1,2 @@
+<h1>Downloads#update</h1>
+<p>Find me in app/views/downloads/update.html.erb</p>
