+++ app/views/layouts/application.html.erb
@@ -20,7 +20,7 @@
   	%>
     <% if current_user.internal_user? %> |
   	<a href="/series">Data Series</a> |
-  	<a href="/data_source_downloads">Downloads</a> |
+  	<a href="/downloads">Downloads</a> |
     <a href="/investigate_visual">Investigation Dashboard</a> |
     <a href="/data_lists">Data Lists</a> |
     <a href="/exports">Exports</a> |
