+++ db/migrate/220170413025719_new_downloads_and_bridge.rb
@@ -1,0 +1,16 @@
+class NewDownloadsAndBridge < ActiveRecord::Migration
+  def up
+    rename_table :data_source_downloads, :downloads
+    create_table :data_source_downloads do |t|
+      t.belongs_to :data_source
+      t.belongs_to :download
+      t.string :last_eval_checksum
+      t.string :last_file_checksum
+    end
+    add_index :data_source_downloads, [:data_source_id, :download_id], unique: true
+  end
+  def down
+    drop_table :data_source_downloads
+    rename_table :downloads, :data_source_downloads
+  end
+end
